Add files via upload
Written in PHP for the backend with some AJAX, it is lacking some regex conditions for the password as I ran out of time and no grades were going for it.  Vanilla JS was used for the FrontEnd and CSS grid for the layout.  When I complete this course I intend to incorporate more AJAX, I don't like how currently when submitting a score the entire screen refreshes and there is a little issue with the welcome back/hello message, but I'll fix that later.  The JSON files work fine and so do cookies, I learned loads about JS alone in this project.